requirements.txt
# FastAPI Framework
fastapi==0.65.2
fastapi-profiler==1.0.0
fastapi-utils==0.2.1
pydantic==1.8.2
uvicorn==0.14.0
starlette==0.14.2
starlette-context==0.3.2

# MTG SDK
mtgsdk==1.3.1

# database migrations
alembic==1.5.8

# Formatting & linting
black==21.4b2
flake8
reorder-python-imports

# Environment Management and Testing
python-dotenv==0.14.0
pytest~=6.0.1
